chore: move new function to a separate package
